requirements.txt
# 🚀 뷰티 클리닉 AI 챗봇 - 핵심 의존성 목록 (Cloud Run 최적화)

# 기본 설정 관리
pydantic-settings>=2.9.1
pydantic>=2.7.0
python-dotenv>=0.21.0

# 웹 프레임워크
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.0

# Google Cloud 핵심 서비스
google-cloud-firestore>=2.13.0
google-cloud-aiplatform>=1.36.0
firebase-admin>=6.2.0

# HTTP 클라이언트
requests>=2.31.0
httpx>=0.25.0

# 날짜/시간 처리
python-dateutil>=2.8.2

# 기본 유틸리티
rich>=13.6.0
